Add timestamp fields to Vendor, Sweet, and VendorSweet models
Add 'created_at' and 'updated_at' fields to the Vendor, Sweet, and VendorSweet models.

Set the default value of 'created_at' and 'updated_at' to the EAT timezone

Update 'updated_at' field automatically whenever the record is updated. - This is achieved by using the 'onupdate' attribute of the SQLAlchemy 'Column' function.